# perf: use direct array access instead of `endsWIth`