--- a/src/storage.ts
+++ b/src/storage.ts
@@ -231,7 +231,7 @@ export function createStorage<T extends StorageValue>(
     async setItems(items, commonOptions) {
       await runBatch(items, commonOptions, async (batch) => {
         if (batch.driver.setItems) {
-          await asyncCall(
+          return asyncCall(
             batch.driver.setItems,
             batch.items.map((item) => ({
               key: item.relativeKey,
@@ -340,10 +340,12 @@ export function createStorage<T extends StorageValue>(
           mount.relativeBase,
           opts
         );
-        const keys = rawKeys
-          .map((key) => mount.mountpoint + normalizeKey(key))
-          .filter((key) => !maskedMounts.some((p) => key.startsWith(p)));
-        allKeys.push(...keys);
+        for (const key of rawKeys) {
+          const fullKey = mount.mountpoint + normalizeKey(key);
+          if (!maskedMounts.some((p) => fullKey.startsWith(p))) {
+            allKeys.push(fullKey);
+          }
+        }
 
         // When /mnt/foo is processed, any key in /mnt with /mnt/foo prefix should be masked
         // Using filter to improve performance. /mnt mask already covers /mnt/foo
@@ -353,8 +355,10 @@ export function createStorage<T extends StorageValue>(
         ];
       }
       return base
-        ? allKeys.filter((key) => key.startsWith(base!) && key[key.length-1] !== "$")
-        : allKeys.filter((key) => key[key.length-1] !== "$");
+        ? allKeys.filter(
+            (key) => key.startsWith(base!) && key[key.length - 1] !== "$"
+          )
+        : allKeys.filter((key) => key[key.length - 1] !== "$");
     },
     // Utils
     async clear(base, opts = {}) {
@@ -412,7 +416,7 @@ export function createStorage<T extends StorageValue>(
           .then((unwatcher) => {
             context.unwatch[base] = unwatcher;
           })
-          .catch(console.error); // eslint-disable-line no-console
+          .catch(console.error);
       }
       return storage;
     },
@@ -447,6 +451,13 @@ export function createStorage<T extends StorageValue>(
         base: m.mountpoint,
       }));
     },
+    // Aliases
+    keys: (base, opts = {}) => storage.getKeys(base, opts),
+    get: (key, opts = {}) => storage.getItem(key, opts),
+    set: (key, value, opts = {}) => storage.setItem(key, value, opts),
+    has: (key, opts = {}) => storage.hasItem(key, opts),
+    del: (key, opts = {}) => storage.removeItem(key, opts),
+    remove: (key, opts = {}) => storage.removeItem(key, opts),
   };
 
   return storage;