@pi0 2024-05-01
Thanks for the PR but I would appreciate sharing a benchmark diff.