@pi0 2024-05-14
Script above is not related to unstoage.

I understand that using index access vs `endsWith` might lead to faster benchmarks but either we need to ban using a javascript feature entirely or only avoid it in critical paths that proofs meaningful performance improvement and for this reason and explicitly mark those lines, for that we need benchmark script first 🙏🏼