@GalacticHypernova 2024-05-14
> either we need to ban using a javascript feature entirely or only avoid it in critical paths that proofs meaningful performance improvement

I mean, I get your point, but I'd have to politely disagree with you on this one. It's not about banning a javascript feature. There are many builtin methods that are there because they have legitimate use cases, but it doesn't necessarily mean that every **possible** use case is an intended use case. I'm sure you've heard about `Array.prototype.reduce`, and the entire section MDN has on it where they show examples of when *not* to use it. It doesn't mean reduce should be banned altogether,  it just means it should be used wherever its actually necessary. And in my opinion this is true for a lot of methods, even endsWith. As with any abstraction, it has its performance considerations, and there are simply cases when using it is not optimal. You might argue that abstractions make the code more readable, but in this scenario I doubt anyone would say that having a simple `if (a[b] === c)` check is unreadable enough to warrant endsWith, and if it doesn't have performance benefits nor readability benefits, I don't see a reason to use it.
It's up to you whether to include this or not, but in such a critical component in the unjs ecosystem, in a part that is required to be as performant as possible, especially with certain drivers, I don't really see a reason not to include this, but I'll the decision up to you.