@GalacticHypernova 2024-05-15
> we need benchmark script first

Do you mean like an actual benchmark test with some benchmarking library? If we're talking about simply showing that it's indeed changing something I have provided a benchmark sample above. EndsWith doesn't discriminate between whether the code is for nuxt or unstorage, the slowdown happens all the time and can easily be observed based on the results that I have also attached, as that's how it was designed. However, if you'd like a full fledged benchmark, then we should already add benchmarks to everything, as nothing is currently benchmarked