@pi0 2024-05-15
Added bench scripts partially specific to this change via https://github.com/unjs/unstorage/commit/d84bcc66d80c1011d7debddfce551afbaecccab4 

You can try with `pnpm bench` and see impl in `test/server.bench.ts` file.

---

Main:

```
storage server     310 µs/iter     (204 µs … 2'470 µs)    337 µs  1'002 µs  1'758 µs
```

This branch:

```
storage server     309 µs/iter     (192 µs … 1'769 µs)    334 µs  1'005 µs  1'666 µs
```

As you can see, the end-to-end difference is almost nothing. Not that it does not matter but hope you understand why micro-optimizations like this (even though, at micro-level are visible) do not essentially improve final performances that much. 

---

I again appreciate your contribution but ask you kindly next times respect maintainer's decision and comments and try to collaborate. Working on the bench took over an hour of my early morning that desperately needed for other stuff.