@GalacticHypernova 2024-05-15
I appreciate the merge! 
I had no problem making the benchmark tests myself, I didn't mean you need to make it, I was simply asking whether it would be good idea to add benchmarks to the project so that I could proceed with it without you having to spend your time on it too much.

I know these are micro optimizations, I never said it would be anything ground breaking, but micro optimizations will eventually add up, especially in performance-critical environments and in I/O bound work.

I don't recall ever refusing to collaborate, I did leave the final decision to you, and I believe I have stated that, and I quote, "It's up to you whether to include this or not" and "I'll leave the decision up to you". I believe there was some miscommunication on the matter, which I would be more than willing to discuss in private if you're willing to.